pyproject.toml
[tool.poetry]
name = "twitter-list-manager"
version = "1.0.0"
description = ""
authors = ["Sakurai <K.Sakurai_2000@outlook.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "3.9.6"
tweepy = "^3.10.0"
pandas = "^1.3.2"
python-dotenv = "^0.19.0"


[tool.poetry.dev-dependencies]
Sphinx = "^4.2.0"
sphinx-rtd-theme = "^1.0.0"
black = "^21.9b0"
flake8 = "^3.9.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
